+++ src/pane.js
@@ -138,18 +138,11 @@ export class Pane extends Cell {
                 return toDo
             })
             this.t.onData(d =>  {
-
-                if (!this.d) {
-                    this.gate.notify("Gate is disconnected")
-                    return
-                }
-                const state = this.d.readyState 
-                if (state != "open") {
-                    this.gate.notify("data channel not ready")
-                    // TODO: maybe try and reconnect?
-                    return
-                }
-                this.d.send(d)
+                if (!this.d || this.d.readyState != "open" ) {
+                    this.gate.notify("Data channel lost")
+                    this.gate.reset()
+                } else
+                    this.d.send(d)
             })
             this.resizeObserver.observe(this.e);
             this.fit(pane => { 
@@ -448,7 +441,7 @@ export class Pane extends Cell {
             break
         // this key is at terminal level
         case "l":
-            f = () => this.t7.logDisplay()
+            f = () => this.t7.map.showLog()
             break
         case "ArrowLeft":
             f = () => this.w.moveFocus("left")